requirements.txt
streamlit
langchain
chromadb
openai
groq
pypdf
